[refactor] 회사 조직도 페이지 일부 컴포넌트화
